PLAT-1308 Turn IOcrTextExtractor into IPdfTextExtractor
- Previously, the interface was quite useless.
- The only implementation so far will only work for PDFs.
- There's an upcoming alternate, PDF-specific implementation in a derived project.
